+++ types/event_bus.go
@@ -115,13 +115,13 @@ func (*EventBus) validateAndStringifyEvents(events []types.Event) map[string][]s
 		if len(event.Type) == 0 {
 			continue
 		}
-
+		prefix := event.Type + "."
 		for _, attr := range event.Attributes {
 			if len(attr.Key) == 0 {
 				continue
 			}
 
-			compositeTag := fmt.Sprintf("%s.%s", event.Type, attr.Key)
+			compositeTag := prefix + attr.Key
 			result[compositeTag] = append(result[compositeTag], attr.Value)
 		}
 	}